add types folder
