+++ app/(root)/users/[id]/page.tsx
@@ -33,29 +33,6 @@ export async function generateMetadata(
   };
 }
 
-export type ArticleWithSaveStatus = {
-  _id: string;
-  title: string;
-  slug: {
-    _type: 'slug';
-    current: string;
-  };
-  _createdAt: string;
-  author: {
-    _id: string;
-    name: string;
-    image: string;
-    bio: string;
-  };
-  views: number;
-  desc: string;
-  category: string;
-  image: string;
-  likes: number;
-  likedBy: string[];
-  saveStatus: boolean;
-};
-
 const UserPage = async ({ params }: { params: Promise<{ id: string }> }) => {
   const id = (await params).id;
   const session = await auth();
