+++ components/Buttons/SaveButton.tsx
@@ -1,9 +1,8 @@
 'use client';
 
 import { useTransition } from 'react';
-import { useRouter } from 'next/navigation';
 import { BookmarkIcon } from 'lucide-react';
-import { Button } from './ui/button';
+import { Button } from '../ui/button';
 import { saveArticle } from '@/actions/server-actions';
 import { toast } from 'sonner';
 import { sendGTMEvent } from '@next/third-parties/google';
